python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "imageio>=2.37.0",
    "matplotlib>=3.10.1",
    "numpy>=2.2.3",
    "pyyaml>=6.0.2",
]
